split NewtonsofSerializer into two classes; NewtonsoftJsonSerializer and NewtonsoftBsonSerializer
