+++ Libraries/RosBridgeClient/RosBridgeClient.csproj
@@ -301,7 +301,8 @@
     <Compile Include="Protocols\WebSocketNetProtocol.cs" />
     <Compile Include="Serializers\ISerializer.cs" />
     <Compile Include="Serializers\MicrosoftSerializer.cs" />
-    <Compile Include="Serializers\NewtonsoftSerializer.cs" />
+    <Compile Include="Serializers\NewtonsoftBsonSerializer.cs" />
+    <Compile Include="Serializers\NewtonsoftJsonSerializer.cs" />
     <Compile Include="UrdfTransfer\UrdfTransfer.cs" />
     <Compile Include="UrdfTransfer\UrdfTransferFromRos.cs" />
     <Compile Include="UrdfTransfer\UrdfTransferToRos.cs" />
